+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IssuingBillingTransaction resource
 - IssuingTokenDesign resource
 - IssuingTokenRequest sub-resource
+- IssuingToken resource
 
 ## [1.1.0] - 2026-05-18
 ### Added